refactor: persistence repository
