+++ src/main/java/fr/insee/pogues/persistence/repository/QuestionnaireRepository.java
@@ -5,30 +5,19 @@ import com.fasterxml.jackson.databind.JsonNode;
 import java.util.List;
 
 /**
- * Questionnaire Service Query interface to assume the persistance of Pogues UI in JSON
- *
- * @author I6VWID
+ * Questionnaire Service Query interface to assume the persistance of Pogues UI in JSON.
  */
-
 public interface QuestionnaireRepository {
 
     JsonNode getQuestionnaireByID(String id) throws Exception;
 
-    JsonNode getJsonLunaticByID(String id) throws Exception;
-
     void deleteQuestionnaireByID(String id) throws Exception;
-    
-    void deleteJsonLunaticByID(String id) throws Exception;
 
     List<JsonNode> getQuestionnairesByOwner(String owner) throws Exception;
 
     void createQuestionnaire(JsonNode questionnaire) throws Exception;
-    
-    void createJsonLunatic(JsonNode questionnaireLunatic) throws Exception;
 
     void updateQuestionnaire(String id, JsonNode questionnaire) throws Exception;
-    
-    void updateJsonLunatic(String id, JsonNode questionnaireLunatic) throws Exception;
 
     List<JsonNode> getMetaQuestionnaire(String owner) throws Exception;
