+++ src/test/java/fr/insee/pogues/service/NomenclatureServiceTest.java
@@ -1,7 +1,7 @@
 package fr.insee.pogues.service;
 
 import fr.insee.pogues.model.*;
-import fr.insee.pogues.persistence.service.QuestionnairesService;
+import fr.insee.pogues.persistence.service.QuestionnaireService;
 import fr.insee.pogues.persistence.service.VersionService;
 import fr.insee.pogues.webservice.model.dtd.nomenclatures.ExtendedNomenclature;
 import org.junit.jupiter.api.BeforeEach;
@@ -20,14 +20,14 @@ import static org.mockito.MockitoAnnotations.initMocks;
 class NomenclatureServiceTest {
 
     @Mock
-    QuestionnairesService questionnairesService;
+    QuestionnaireService questionnaireService;
     @Mock
     VersionService versionService;
     private NomenclatureService nomenclatureService;
 
     @BeforeEach
     void initQuestionnaireService(){
-        nomenclatureService = new NomenclatureService(questionnairesService, versionService);
+        nomenclatureService = new NomenclatureService(questionnaireService, versionService);
         initMocks(this);
     }
